refactoring mappers and directories, move resoruce result to common module insted domain module
