+++ phrasalito/phrasalito_data/src/main/java/com/nayibit/phrasalito_data/entities/DeckEntity.kt
@@ -10,11 +10,6 @@ data class DeckEntity(
     val id: Int = 0,
     val name: String,
     val maxCards: Int
-){
-    fun toDomain(): Deck {
-        return Deck(
-            id = id,
-            name = name,
-            maxCards = maxCards)
-    }
-}
+)
+
+
