+++ phrasalito/phrasalito_data/src/main/java/com/nayibit/phrasalito_data/repository/DeckRepositoryImpl.kt
@@ -1,16 +1,15 @@
 package com.nayibit.phrasalito_data.repository
 
+import com.nayibit.common.util.Resource
 import com.nayibit.phrasalito_data.dao.DeckDao
-import com.nayibit.phrasalito_data.entities.DeckEntity
+import com.nayibit.phrasalito_data.mapper.toDomain
+import com.nayibit.phrasalito_data.mapper.toEntity
 import com.nayibit.phrasalito_domain.model.Deck
 import com.nayibit.phrasalito_domain.repository.DeckRepository
-import com.nayibit.phrasalito_domain.utils.Resource
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.map
 import javax.inject.Inject
-import kotlin.collections.map
 
 class DeckRepositoryImpl
     @Inject constructor(private val deckDao: DeckDao) : DeckRepository  {
@@ -19,11 +18,7 @@ class DeckRepositoryImpl
             emit(Resource.Loading)
             delay(1000)
          try {
-             val deckEntity = DeckEntity(
-                 name = deck.name,
-                 maxCards = deck.maxCards
-             )
-             deckDao.insert(deckEntity)
+             deckDao.insert(deck.toEntity())
              emit(Resource.Success(deck))
          }catch (e: Exception){
             emit(Resource.Error(e.localizedMessage ?: "Unknown error"))
