+++ phrasalito/phrasalito_domain/src/main/java/com/nayibit/phrasalito_domain/useCases/decks/GetAllDecksUseCase.kt
@@ -1,8 +1,8 @@
 package com.nayibit.phrasalito_domain.useCases.decks
 
+import com.nayibit.common.util.Resource
 import com.nayibit.phrasalito_domain.model.Deck
 import com.nayibit.phrasalito_domain.repository.DeckRepository
-import com.nayibit.phrasalito_domain.utils.Resource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
