+++ phrasalito/phrasalito_domain/src/main/java/com/nayibit/phrasalito_domain/useCases/phrases/DeletebyIdPhraseUseCase.kt
@@ -1,7 +1,7 @@
 package com.nayibit.phrasalito_domain.useCases.phrases
 
+import com.nayibit.common.util.Resource
 import com.nayibit.phrasalito_domain.repository.PhraseRepository
-import com.nayibit.phrasalito_domain.utils.Resource
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 import javax.inject.Inject
