+++ phrasalito/phrasalito_domain/src/main/java/com/nayibit/phrasalito_domain/useCases/phrases/GetAllPhrasesUseCase.kt
@@ -1,8 +1,8 @@
 package com.nayibit.phrasalito_domain.useCases.phrases
 
+import com.nayibit.common.util.Resource
 import com.nayibit.phrasalito_domain.model.Phrase
 import com.nayibit.phrasalito_domain.repository.PhraseRepository
-import com.nayibit.phrasalito_domain.utils.Resource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
