+++ phrasalito/phrasalito_presentation/src/main/java/com/nayibit/phrasalito_presentation/screens/phraseScreen/PhraseViewModel.kt
@@ -3,13 +3,13 @@ package com.nayibit.phrasalito_presentation.screens.phraseScreen
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.nayibit.common.util.Resource
 import com.nayibit.common.util.UiText
 import com.nayibit.phrasalito_domain.model.Phrase
 import com.nayibit.phrasalito_domain.useCases.phrases.DeletebyIdPhraseUseCase
 import com.nayibit.phrasalito_domain.useCases.phrases.GetAllPhrasesUseCase
 import com.nayibit.phrasalito_domain.useCases.phrases.InsertPhraseUseCase
 import com.nayibit.phrasalito_domain.useCases.phrases.UpdatePhraseByIdUseCase
-import com.nayibit.phrasalito_domain.utils.Resource
 import com.nayibit.phrasalito_presentation.R
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
